many changes, but not 100% working! don't use now
editor:
bricks, floor, pipes: can't change width-property anymore, instead use the blocks one
show function/class constructor signature in title when using it
automatic backup file if the editor crashes
fixed design-bug
fix crash on set direction=None on turtles

game:
added levelpackswitcher if you install some level pack (no levelpack released yet)
delete points when your tux is dead
fixed flyingcircle levelscript
made game object speed's relative to frame rate

To show FPS, press 'f'

+++ littleRunner/Drawing/Draw/Draw.cs
@@ -8,10 +8,10 @@ namespace littleRunner.Drawing
 {
     abstract public class Draw
     {
-        public abstract void DrawImage(dImage image, int x, int y, int width, int height);
-        public abstract void DrawRectangle(dPen pen, int x, int y, int width, int height);
-        public abstract void FillRectangle(dPen pen, int x, int y, int width, int height);
-        public abstract void DrawString(string text, dFont font, dColor color, int x, int y);
-        public abstract void MoveCoords(int x, int y);
+        public abstract void DrawImage(dImage image, float x, float y, int width, int height);
+        public abstract void DrawRectangle(dPen pen, float x, float y, int width, int height);
+        public abstract void FillRectangle(dPen pen, float x, float y, int width, int height);
+        public abstract void DrawString(string text, dFont font, dColor color, float x, float y);
+        public abstract void MoveCoords(float x, float y);
     }
 }
